feat($goKey): Sync local object w. GoInstant key
The $goKey service has been introduced, it takes a single argument, a
key.  $goKey returns a model, the standard `key` methods are exposed,
prefixed with a '$'. This model will be kept current with the key.

Note: Changes to the object are not persisted back to the key.

For example, to return a collection of todos:

$scope.todos = $goKey(lobby.key('todos'));
$scope.todos.$sync();

To add a new todo:

$scope.todos.$add({
  description: 'a fresh todo',
  complete: false
});

Each of the methods with the exception of $sync, returns a promise.

Additional minor changes:

Also added grunt-contrib-watch, it adds `grunt watch`, which:
- monitors the lib folder and runs component build -d
- monitors the build file and live-reloads

Closes #28, #23